plugins/skills/orchestrate-integrate/requirements.txt
rich>=13.7.0
typer>=0.9.0
pydantic>=2.5.0
pyyaml>=6.0.1
